+++ plugins/funind/functional_principles_proofs.ml
@@ -995,7 +995,7 @@ let generate_equation_lemma evd fnames f fun_num nb_params nb_args rec_args_num 
       Ensures by: obvious
       i*)
     (mk_equation_id f_id)
-    (Decl_kinds.Global, false, (Decl_kinds.Proof Decl_kinds.Theorem))
+    Decl_kinds.(Global ImportDefaultBehavior, false, Proof Theorem)
     evd
   lemma_type
   in