+++ dev/doc/changes.md
@@ -5,6 +5,10 @@
 - Functions and types deprecated in 8.10 have been removed in Coq
   8.11.
 
+- Type Decl_kinds.locality has been restructured, see commit
+  message. Main change to do generally is to change the flag "Global"
+  to "Global ImportDefaultBehavior".
+
 ## Changes between Coq 8.9 and Coq 8.10
 
 ### ML4 Pre Processing